feat: allow configuring server.properties
Configuring the server.properties file is managed by passing environment
variables prefixed with `MC_`. The prefix is removed, any `_` are
replaced with `-` and the value is lowercased to find the correct
server.properties key.

For example: `MC_MAX_PLAYERS` becomes `max-players`

refactor: separate user configuration to a stand-alone script

This will allow this script to be moved elsewhere to simplify
maintenance and allow for reusability.

docs: create readme file